+++ example-project/mycmsplugin/plugins.py
@@ -1,0 +1,6 @@
+from mycmsproject.plugins import ContentType
+
+
+class BlogPost(ContentType):
+    title = 'Blog post'
+    name = 'blog-post'