+++ example-project/mycmsproject/forms.py
@@ -1,0 +1,8 @@
+from django import forms
+
+from .models import Content
+
+
+class ContentForm(forms.ModelForm):
+    class Meta:
+        model = Content