+++ example-project/mycmsproject/models.py
@@ -1,0 +1,14 @@
+from django.db import models
+
+from djangoplugins.fields import PluginField
+
+from .plugins import ContentType
+
+
+class Content(models.Model):
+    title = models.CharField(max_length=255)
+    content = models.TextField()
+    plugin = PluginField(ContentType, editable=False)
+
+    def get_absolute_url(self):
+        return self.plugin.get_plugin().get_read_url(self)